refactor(YearsWithMultipleWinnersTable): extract to its own folder
