+++ frontend/src/utils/constants.js
@@ -1,0 +1,1 @@
+export const BASE_API_URL = 'https://tools.texoit.com/backend-java/api/movies';
